Update custom.theme.css
Changed background image to use an embedded base64 version instead of downloading it on program startup.  This saves loading time although does inflate the css size itself.  Helps for those of us on slower connections...  I'm keeping a commented version of the line in the file to show the source url.  It will be updated if/when the image is changed.